Moved util function and types out of hook file into util file [RCORE-647]
